requirements.txt
requests
beautifulsoup4
lxml
markdownify
python-frontmatter
python-dotenv
feedparser
tqdm
pdfminer.six
python-frontmatter
faiss-cpu
sentence-transformers